feat: 장목 금수 판단 객체 추가
