+++ src/test/kotlin/omok/model/rule/count/OverlineRowRuleTest.kt
@@ -1,0 +1,3 @@
+package omok.model.rule.count
+
+class OverlineRowRuleTest
